requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
mysqlclient==2.2.6
python-dateutil==2.9.0.post0
Pillow==11.0.0
coreapi==2.3.3
pygments==2.18.0
markdown==3.7
